Do not misprint large TCP info values
Previously, values like snd_(c)wnd were stored as int in struct
iperf_interval_results, while being handled as long before and double
after. This led to large values being wrongly printed as negative
values, e.g. when the product of snd_cwnd and tcpi_snd_mss in
get_snd_cwnd() resulted in a large return value.

This change is still not without issues. Ideally, the products of two
uint32_ts in get_snd_wnd() and get_snd_cwnd() would be returned as a
uint64_t. But that would make -1 unavailable as a placeholder for
missing values.

+++ src/iperf.h
@@ -101,14 +101,14 @@ struct iperf_interval_results
     /* Just placeholders, never accessed. */
     char *tcpInfo;
 #endif
-    int interval_retrans;
-    int snd_cwnd;
-    int snd_wnd;
+    long interval_retrans;
+    long snd_cwnd;
+    long snd_wnd;
     TAILQ_ENTRY(iperf_interval_results) irlistentries;
     void     *custom_data;
-    int rtt;
-    int rttvar;
-    int pmtu;
+    long rtt;
+    long rttvar;
+    long pmtu;
 };
 
 struct iperf_stream_result
@@ -118,14 +118,14 @@ struct iperf_stream_result
     iperf_size_t bytes_received_this_interval;
     iperf_size_t bytes_sent_this_interval;
     iperf_size_t bytes_sent_omit;
-    int stream_prev_total_retrans;
-    int stream_retrans;
-    int stream_max_rtt;
-    int stream_min_rtt;
-    int stream_sum_rtt;
+    long stream_prev_total_retrans;
+    long stream_retrans;
+    long stream_max_rtt;
+    long stream_min_rtt;
+    long stream_sum_rtt;
     int stream_count_rtt;
-    int stream_max_snd_cwnd;
-    int stream_max_snd_wnd;
+    long stream_max_snd_cwnd;
+    long stream_max_snd_wnd;
     struct iperf_time start_time;
     struct iperf_time end_time;
     struct iperf_time start_time_fixed;
